ml/requirements.txt
matplotlib==3.9.2
mplfinance==0.12.10b0
pandas==2.2.3
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
yfinance==0.2.54
